Hide jump bar if you have no folders
